Feature: supplement incomplete import records
When an import record is "incomplete" per #9440, this commit will look
to the `import_item` table to supplement that `rec` with data loaded by
BookWorm.

This attempt to supplement will apply to any import source if one of the
required-for-completeness fields (title, authors, publish_date) is
missing.

See https://github.com/internetarchive/openlibrary/issues/9440.
