+++ openlibrary/catalog/add_book/__init__.py
@@ -24,7 +24,6 @@ A record is loaded by calling the load function.
 """
 
 import itertools
-import json
 import re
 from typing import TYPE_CHECKING, Any, Final
 
@@ -982,32 +981,6 @@ def should_overwrite_promise_item(
     return bool(safeget(lambda: edition['source_records'][0], '').startswith("promise"))
 
 
-def supplement_rec_with_import_item_metadata(
-    rec: dict[str, Any], identifier: str
-) -> None:
-    """
-    Queries for a staged/pending row in `import_item` by identifier, and if found, uses
-    select metadata to supplement empty fields/'????' fields in `rec`.
-
-    Changes `rec` in place.
-    """
-    from openlibrary.core.imports import ImportItem  # Evade circular import.
-
-    import_fields = [
-        'authors',
-        'publish_date',
-        'publishers',
-        'number_of_pages',
-        'physical_format',
-    ]
-
-    if import_item := ImportItem.find_staged_or_pending([identifier]).first():
-        import_item_metadata = json.loads(import_item.get("data", '{}'))
-        for field in import_fields:
-            if not rec.get(field) and (staged_field := import_item_metadata.get(field)):
-                rec[field] = staged_field
-
-
 def load(rec: dict, account_key=None, from_marc_record: bool = False):
     """Given a record, tries to add/match that edition in the system.
 
@@ -1026,10 +999,6 @@ def load(rec: dict, account_key=None, from_marc_record: bool = False):
         validate_record(rec)
 
     normalize_import_record(rec)
-
-    # For recs with a non-ISBN ASIN, supplement the record with BookWorm metadata.
-    if non_isbn_asin := get_non_isbn_asin(rec):
-        supplement_rec_with_import_item_metadata(rec=rec, identifier=non_isbn_asin)
 
     # Resolve an edition if possible, or create and return one if not.
     edition_pool = build_pool(rec)
